+++ Data/src/main/java/ru/jsavings/data/model/database/transaction/Transaction.kt
@@ -1,6 +1,5 @@
 package ru.jsavings.data.model.database.transaction
 
-import ru.jsavings.data.model.common.BaseModel
 import ru.jsavings.data.model.database.common.BaseDbModel
 import ru.jsavings.data.model.database.purse.Purse
 import java.util.*
@@ -8,7 +7,7 @@ import java.util.*
 data class Transaction (
 	val transactionId: Long = 0,
 	val purse: Purse,
-	val transactionCategoryId: Int,
+	val transactionCategoryId: Long,
 	val totalSum: Double,
 	val date: Date,
 	val description: String = "",
