+++ Data/src/main/java/ru/jsavings/data/model/database/transaction/TransactionCategory.kt
@@ -3,7 +3,7 @@ package ru.jsavings.data.model.database.transaction
 import ru.jsavings.data.model.database.common.BaseDbModel
 
 data class TransactionCategory (
-	val categoryId: Int = 0,
+	val categoryId: Long = 0,
 	val name: String,
 	val accountId: Int,
 	val categoryType: TransactionCategoryType,
